pyproject.toml
[tool.poetry]
name = "mpd-cast"
version = "0.1.0"
description = "MPD server for Chromecasts"
authors = ["David Hoyes <dphoyes@gmail.com>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.11"
anyio = "^4.8.0"
trio = "^0.28.0"
PyChromecast = "^10.3.0"
Flask = "^2.0.3"
python-mpd-server = {git = "git@github.com:dphoyes/python-mpd-server.git", branch = "master", develop = true}

[tool.poetry.scripts]
mpd-cast = 'mpd_cast.mpd:MainProgram.main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
